backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
openai==1.3.0
assemblyai==0.21.0
runwayml==0.1.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-flask==1.3.0 